requirements.txt
requests==2.22.0
numpy==1.19.3
scikit-learn==0.24.2
google-cloud-storage==1.17.0
Flask>=2.2.2<3.1.0
gunicorn